moves withMenuState to utils
